Add department interface and move types to shared folder
