+++ src/modules/details/actions/getArtwork.ts
@@ -1,14 +1,14 @@
 'use server'
 
-import type { Artwork } from '~/modules/artworks/types'
 import { API } from '~/shared/helpers/api'
+import type { Artwork } from '~/shared/types'
 
 interface GetArtworkPayload {
   objectID: number
 }
 
 export async function getArtwork({ objectID }: GetArtworkPayload) {
-  const { data: artwork } = await API.get<Artwork>(`/objects/${objectID}`)
+  const { data } = await API.get<Artwork>(`/objects/${objectID}`)
 
-  return artwork
+  return data
 }
